+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2
 - Add `Bookmarks: Go to Next in Current Editor` and `Bookmarks: Go to Previous in Current Editor` commands.
+- Add `Remove All` command to context menu of `Workspace` and `Global` tree nodes.
 ## 0.1.0/0.1.1
 - Non-preview release
 - Set `contextualTitle` so explorer tooltip is `Bookmarks`, not `Explorer`.